+++ hbw/app/javascript/packs/hbw/components/form/group.js.jsx
@@ -3,7 +3,7 @@ import compose from 'shared/utils/compose';
 import { withConditions, withErrorBoundary } from 'shared/hoc';
 import TranslationContext from 'shared/context/translation';
 import FileList from './file_list';
-import ServicesTable from './services_table';
+import ServicesTable from './services_table/services_table';
 
 modulejs.define(
   'HBWFormGroup',