+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v2.9 [Unreleased]
 ### Features
 - [#730](https://github.com/hydra-billing/homs/pull/730) Update dependencies.
+- [#729](https://github.com/hydra-billing/homs/pull/729/) Add services select table search option.
 - [#726](https://github.com/hydra-billing/homs/pull/726/) Add services select table.
 - [#725](https://github.com/latera/homs/pull/725) Update Rails to 7.
 - [#706](https://github.com/latera/homs/pull/706) Update Ruby to 3.2.2.